requirements.txt
asgiref==3.5.2
Django==4.1.2
django-debug-toolbar==3.7.0
django-decouple==2.1
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2-binary==2.9.4
pytz==2022.5
sqlparse==0.4.3
